Refactor printer into own base dir
