+++ InvenTree/plugin/base/event/events.py
@@ -14,8 +14,6 @@ from django.db import transaction
 from django.db.models.signals import post_save, post_delete
 from django.dispatch.dispatcher import receiver
 
-import common.notifications
-
 from InvenTree.ready import canAppAccessDatabase, isImportingData
 from InvenTree.tasks import offload_task
 
@@ -194,46 +192,3 @@ def after_delete(sender, instance, **kwargs):
         f'{table}.deleted',
         model=sender.__name__,
     )
-
-
-def print_label(plugin_slug, label_image, label_instance=None, user=None):
-    """
-    Print label with the provided plugin.
-
-    This task is nominally handled by the background worker.
-
-    If the printing fails (throws an exception) then the user is notified.
-
-    Arguments:
-        plugin_slug: The unique slug (key) of the plugin
-        label_image: A PIL.Image image object to be printed
-    """
-
-    logger.info(f"Plugin '{plugin_slug}' is printing a label")
-
-    plugin = registry.plugins.get(plugin_slug, None)
-
-    if plugin is None:
-        logger.error(f"Could not find matching plugin for '{plugin_slug}'")
-        return
-
-    try:
-        plugin.print_label(label_image, width=label_instance.width, height=label_instance.height)
-    except Exception as e:
-        # Plugin threw an error - notify the user who attempted to print
-
-        ctx = {
-            'name': _('Label printing failed'),
-            'message': str(e),
-        }
-
-        logger.error(f"Label printing failed: Sending notification to user '{user}'")
-
-        # Throw an error against the plugin instance
-        common.notifications.trigger_notifaction(
-            plugin.plugin_config(),
-            'label.printing_failed',
-            targets=[user],
-            context=ctx,
-            delivery_methods=[common.notifications.UIMessageNotification]
-        )
