+++ InvenTree/plugin/mixins/__init__.py
@@ -2,13 +2,14 @@
 Utility class to enable simpler imports
 """
 
-from ..base.integration.mixins import APICallMixin, AppMixin, LabelPrintingMixin, SettingsMixin, ScheduleMixin, UrlsMixin, NavigationMixin, PanelMixin
+from ..base.integration.mixins import APICallMixin, AppMixin, SettingsMixin, ScheduleMixin, UrlsMixin, NavigationMixin, PanelMixin
 
 from common.notifications import SingleNotificationMethod, BulkNotificationMethod
 
 from ..base.action.mixins import ActionMixin
 from ..base.barcodes.mixins import BarcodeMixin
 from ..base.event.mixins import EventMixin
+from ..base.label.mixins import LabelPrintingMixin
 
 __all__ = [
     'APICallMixin',
